Fix dirty degree evaluation
We used to simply add up all values in non-obstacle cells, while GLUT
version summed squares of cells' values. As of now we do the same.

+++ world.h
@@ -112,8 +112,8 @@ private:
     Agent agent;
 
     int currentTime;
-    int dirtyDegree;
-    int consumedEnergy;
+    unsigned long dirtyDegree;
+    unsigned long consumedEnergy;
 
     Agent::actions lastAgentAction;
